requirements.txt
openpyxl==3.0.10
pandas==1.5.3
numpy==1.24.1
tqdm==4.64.1
python-dotenv==0.21.1